feat: create new ReferenceError exception type fix: dont remove asts for assignments to literals unless both sides are marked as literals chore: move exceptions into their own header files, into an exceptions directory
