+++ src/exceptions/Error.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+using namespace std;
+
+namespace Theta {
+    class Error : public exception {
+        public:
+            virtual void display() = 0;
+    };
+}
+
