--- src/util/exceptions.hpp
@@ -1,160 +1,0 @@
-#ifndef EXCEPTIONS_H
-#define EXCEPTIONS_H
-
-#include "compiler/TypeChecker.hpp"
-#include <memory>
-#import <string>
-#include <utility>
-#import <vector>
-#import <iostream>
-#import "../parser/ast/TypeDeclarationNode.hpp"
-#import "../lexer/Token.hpp"
-
-using namespace std;
-
-namespace Theta {
-    class Error : public exception {
-        public:
-            virtual void display() = 0;
-    };
-
-    class CompilationError : public Error {
-        private:
-            string errorType;
-            string message;
-            Token token;
-            string source;
-            string fileName;
-
-        public:
-            CompilationError(string type, string msg, Token tok, string &src, string file) : errorType(type), message(msg), token(tok), source(src), fileName(file) {};
-
-            string what() {
-                return message + " at line " + to_string(token.getStartLocation()[0]) + ", column " + to_string(token.getStartLocation()[1]);
-            }
-
-            void display() override {
-                cout << "\n" + fileName << endl;
-                cout << "  \033[1;31m" + errorType + "\033[0m: " << what() << ':' << endl;
-
-                string contextPrevLine;
-                string contextErrorLine;
-                string contextNextLine;
-                int line = 1;
-                int charIdx = 0;
-
-                // If the error isn't on the first line of the file, we want to add a prevLine to the log output, for context.
-                if (token.getStartLocation()[0] > 1) {
-                    for (; line < token.getStartLocation()[0] - 1; charIdx++) {
-                        if (source[charIdx] == '\n') line++;
-                    }
-
-                    while (source[charIdx] != '\n') {
-                        contextPrevLine += source[charIdx];
-                        charIdx++;
-                    }
-                }
-
-                // Skip to the line where the error is
-                for (; line < token.getStartLocation()[0]; charIdx++) {
-                    if (source[charIdx] == '\n') line++;
-                }
-
-                while (source[charIdx] != '\n' && charIdx <= source.length()) {
-                    contextErrorLine += source[charIdx];
-                    charIdx++;
-                }
-
-                // If there's another line after this one, add its contents to the nextLine for logging output
-                if (charIdx++ < source.length()) {
-                    while (source[charIdx] != '\n') {
-                        contextNextLine += source[charIdx];
-                        charIdx++;
-                    }
-                }
-
-                string errorMarker(token.getStartLocation()[1] + to_string(token.getStartLocation()[0]).length() + 1, ' ');
-                string errorPoint(token.getLexeme().length(), '^');
-
-                if (contextPrevLine != "") {
-                    cout << "    " + to_string(token.getStartLocation()[0] - 1) + ": " + contextPrevLine << endl;
-                }
-
-                cout << "    " + to_string(token.getStartLocation()[0]) + ": " +  contextErrorLine << endl;
-                cout << "    " + errorMarker + "\033[31m" + errorPoint + "\033[0m" << endl;
-
-                if (contextNextLine != "") {
-                    cout << "    " + to_string(token.getStartLocation()[0] + 1) + ": " + contextNextLine << endl;
-                }
-            }
-    };
-
-    class TypeError : public Error {
-        public:
-            TypeError(string msg, shared_ptr<ASTNode> t1, shared_ptr<ASTNode> t2) : message(msg), type1(t1), type2(t2) {}
-
-            string message;
-            shared_ptr<ASTNode> type1;
-            shared_ptr<ASTNode> type2;
-
-            void display() override {
-                string errText = "  \033[1;31mTypeError\033[0m: " + message + ": "; 
-
-                pair<string, string> typeDiff = getTypeDiff(
-                    dynamic_pointer_cast<TypeDeclarationNode>(type1),
-                    dynamic_pointer_cast<TypeDeclarationNode>(type2)
-                );
-
-                if (typeDiff.first == "VARIADIC_MISMATCH") {
-                    errText += "Variadic right hand side must satisfy at least one left hand side type, and may not include extra types";
-                } else {
-                    errText += typeDiff.first + " is not equivalent to " + typeDiff.second;
-                }
-
-                cout << errText << endl;
-            }
-        
-        private:
-            static pair<string, string> getTypeDiff(shared_ptr<TypeDeclarationNode> t1, shared_ptr<TypeDeclarationNode> t2) {
-                if (!t1 || !t2) {
-                    return make_pair((t1 ? t1->getType() : "Nothing"), (t2 ? t2->getType() : "Nothing"));
-                } else if (t1->getType() != t2->getType()) {
-                    return make_pair(
-                        (t1->getType() != "" ? t1->getType() : "Nothing"), 
-                        (t2->getType() != "" ? t2->getType() : "Nothing")
-                    );
-                } else if (t1->getValue() || t2->getValue()) {
-                    return getTypeDiff(
-                        dynamic_pointer_cast<TypeDeclarationNode>(t1->getValue()), 
-                        dynamic_pointer_cast<TypeDeclarationNode>(t2->getValue())
-                    );
-                } else if (t1->getLeft() || t2->getLeft()) {
-                    if (!TypeChecker::isSameType(t1->getLeft(), t2->getLeft())) {
-                        return getTypeDiff(
-                            dynamic_pointer_cast<TypeDeclarationNode>(t1->getLeft()),
-                            dynamic_pointer_cast<TypeDeclarationNode>(t2->getLeft())
-                        );
-                    }
-                    
-                    return getTypeDiff(
-                        dynamic_pointer_cast<TypeDeclarationNode>(t1->getRight()),
-                        dynamic_pointer_cast<TypeDeclarationNode>(t2->getRight())
-                    );
-                } 
-
-                return make_pair("VARIADIC_MISMATCH", "VARIADIC_MISMATCH");
-            }
-    };
-
-    class ParseError : public exception {
-        public:
-            ParseError(string typ) : errorParseType(typ) {}
-
-            string getErrorParseType() { return errorParseType; }
-
-        private:
-            string errorParseType;
-    };
-}
-
-#endif
